# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SHOPIN",
        "description": "SHOPIN API's integration",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get a single category by ID",
                "description": "Returns the details of a single active category with full icon URLs.",
                "operationId": "getCategoryById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "titre": {
                                                    "type": "string",
                                                    "example": "Fashion & Vêtements"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Fashion & Vêtements looollmm"
                                                },
                                                "pourcentage_gain": {
                                                    "type": "string",
                                                    "example": "7.00"
                                                },
                                                "luxury": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "order": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "icon": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/storage/uploads/categories/goCedMgqRvvQQ7xzuwNBgraFZSuuvWShzRCknYvP.png"
                                                },
                                                "small_icon": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/storage/uploads/categories/oE8n20szUVwCTh4pSCy8VXSP0KvTliyMBgX2kDxg.png"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-09-24T11:24:48.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-06-14T17:41:17.000000Z"
                                                },
                                                "active": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title_en": {
                                                    "type": "string",
                                                    "example": "Fashion & Clothing"
                                                },
                                                "title_ar": {
                                                    "type": "string",
                                                    "example": "الموضة والملابس"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Categorie not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}